Update NrcsAwdb.cfg
The server (https://www.wcc.nrcs.usda.gov/) which serves the documentation for the NRCS webservice will be decommissioned likely by the end of Oct 2021, I have swapped the urls with the new locations - sadly these will change again in a year or so, but I'll try to remember to circle back. Note the WSDL will still reside at wcc.sc.egov.usda.gov, that is likely stable for a long time.